feat: rename 'footer' to 'appfooter' because it's specifically for whole apps
