+++ src/components/AppFooter/index.js
@@ -1,0 +1,3 @@
+import AppFooter from "./AppFooter";
+
+export { AppFooter };
