--- src/components/Footer/index.js
@@ -1,3 +1,0 @@
-import Footer from './Footer';
-
-export { Footer }
